feat: add base adapters _statsforecast
